requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.2
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
Werkzeug==3.1.4
